Decoupled paintViewModel and INetworkBroadcaster.

+++ LANPaint_vNext/Services/UDP/INetworkBroadcaster.cs
@@ -6,7 +6,7 @@ namespace LANPaint_vNext.Services.UDP
     public interface INetworkBroadcaster : IDisposable
     {
         Task<byte[]> ReceiveAsync();
-        Task<int> SendAsync(byte[] bytes);
+        Task<long> SendAsync(byte[] bytes);
         ValueTask ClearBufferAsync();
     }
 }
