+++ LANPaint_vNext/Services/UDP/UDPBroadcastBase.cs
@@ -32,7 +32,7 @@ namespace LANPaint_vNext.Services.UDP
             Client = new UdpClient(new IPEndPoint(LocalIp, Port));
         }
 
-        public abstract Task<int> SendAsync(byte[] bytes);
+        public abstract Task<long> SendAsync(byte[] bytes);
         public abstract Task<byte[]> ReceiveAsync();
 
         public virtual async ValueTask ClearBufferAsync()
